rename `query-utility` to `query-filter`
make it clear that only filters are expected and, generally, filters are the way to go for adding queries.
