+++ packages/lix-sdk/src/query-filter/change-conflict-in-version.ts
